+++ skia-safe/src/core/canvas.rs
@@ -2,7 +2,7 @@ use std::ops::{Deref, DerefMut};
 use std::marker::PhantomData;
 use std::slice;
 use std::ffi::CString;
-use crate::graphics;
+use crate::gpu;
 use crate::prelude::*;
 use crate::core::{IRect, QuickReject, Region, RRect, ClipOp, Point, scalar, Vector, Image, ImageFilter, Rect, IPoint, Surface, Bitmap, ISize, SurfaceProps, ImageInfo, Path, Paint, Color, Matrix, BlendMode, Font, TextEncoding, Picture, Vertices, VerticesBone, Data, TextBlob};
 use skia_bindings::{C_SkAutoCanvasRestore_destruct, SkAutoCanvasRestore, C_SkCanvas_isClipEmpty, C_SkCanvas_discard, SkCanvas_PointMode, SkImage, SkImageFilter, SkPaint, SkRect, C_SkCanvas_getBaseLayerSize, C_SkCanvas_imageInfo, C_SkCanvas_newFromBitmapAndProps, C_SkCanvas_newFromBitmap, C_SkCanvas_newWidthHeightAndProps, C_SkCanvas_newEmpty, C_SkCanvas_MakeRasterDirect, SkCanvas, C_SkCanvas_delete, C_SkCanvas_makeSurface, C_SkCanvas_getGrContext, SkCanvas_SaveLayerRec, SkMatrix, SkCanvas_SrcRectConstraint, C_SkAutoCanvasRestore_restore, C_SkAutoCanvasRestore_Construct, SkCanvas_SaveLayerFlagsSet_kInitWithPrevious_SaveLayerFlag};
@@ -286,8 +286,8 @@ impl Canvas {
     }
 
     // TODO: test ref count consistency assuming it is not increased in the native part.
-    pub fn graphics_context(&mut self) -> Option<graphics::Context> {
-        graphics::Context::from_unshared_ptr(unsafe {
+    pub fn graphics_context(&mut self) -> Option<gpu::Context> {
+        gpu::Context::from_unshared_ptr(unsafe {
             C_SkCanvas_getGrContext(self.native_mut())
         })
     }
