+++ skia-safe/src/core/surface.rs
@@ -2,7 +2,7 @@ use std::ptr;
 use std::marker::PhantomData;
 use std::ops::{Deref, DerefMut};
 use crate::prelude::*;
-use crate::graphics::{Context, BackendTexture, BackendRenderTarget, SurfaceOrigin};
+use crate::gpu::{Context, BackendTexture, BackendRenderTarget, SurfaceOrigin};
 use crate::core::{ImageInfo, SurfaceProps, Paint, ColorSpace, Budgeted, IRect, Size, IPoint, Bitmap, Image, Canvas, ISize, ColorType };
 use skia_bindings::{SkSurface, SkSurface_BackendHandleAccess, GrBackendTexture, SkRefCntBase, SkSurface_ContentChangeMode, GrBackendRenderTarget, C_SkSurface_makeSurface};
 #[cfg(test)]
